Added a forgotten cast to randomnumber in var
this is like covering potholes after potholes

+++ AxonCorruptor/NumberEngine.cs
@@ -37,13 +37,13 @@ namespace AxonCorruptor
                         content = File.ReadAllText(filenames[i]);
                         for (int j = 0; j < intensity; j++)
                         {
-                            int randomnumber1 = RandomNumber((int)numericUpDown1.Value, (int)numericUpDown2.Value);
+                            int randomnumber1 = (int)RandomNumber((int)numericUpDown1.Value, (int)numericUpDown2.Value);
                             int foundindex = content.IndexOf(randomnumber1.ToString(),RandomNumber(0,content.Length - 1));
                             if (foundindex != -1) 
                             {
                                 Console.WriteLine("Found the value at index " + foundindex);
                                 Debug.WriteLine("Found the value at index " + foundindex);
-                                int randomnumber2 = RandomNumber((int)numericUpDown1.Value, (int)numericUpDown2.Value);
+                                int randomnumber2 = (int)RandomNumber((int)numericUpDown1.Value, (int)numericUpDown2.Value);
                                 content = content.Remove(foundindex, randomnumber1.ToString()).Insert(foundindex,randomnumber2.ToString());
                                 randomnumber2 = null;
                             }
